Moved source files and:
* created new version files
* added build / test / lint targets in new Makefile
* added Travis CI + git ignore file
