+++ Makefile
@@ -1,0 +1,46 @@
+VERSION_SRC = src/github.com/zylisp/gisp/gitcommit.go
+LAST_TAG = $(shell git describe --abbrev=0 --tags)
+LAST_COMMIT = $(shell git rev-parse --short HEAD)
+
+.PHONY: build test all
+
+all: build test
+
+deps:
+	echo
+
+lint-deps:
+	curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | \
+	sh -s -- -b ~/go/bin v1.15.0
+
+test-deps:
+	go get github.com/masukomi/check
+
+build: deps
+	/bin/echo "package gisp" > $(VERSION_SRC)
+	/bin/echo "" >> $(VERSION_SRC)
+	/bin/echo "func init() { GITLASTTAG = \"$(LAST_TAG)\"; \
+	GITLASTCOMMIT = \"$(LAST_COMMIT)\" }" >> $(VERSION_SRC)
+	go install github.com/zylisp/gisp/cmd/gisp
+
+lint: lint-deps
+	golangci-lint run
+
+vet:
+	go vet github.com/zylisp/gisp/
+	# go vet github.com/zylisp/gisp/cmd
+	# go vet github.com/zylisp/gisp/core
+	# go vet github.com/zylisp/gisp/generator
+	# go vet github.com/zylisp/gisp/lexer
+	# go vet github.com/zylisp/gisp/parser
+
+
+test: test-deps
+	echo "running 'go test' for core ..." && \
+	cd src/github.com/zylisp/gisp/core && \
+	go test -v && \
+	cd - && \
+	echo "running 'go test' for lexer ..." && \
+	cd src/github.com/zylisp/gisp/lexer && \
+	go test -v
+
