+++ src/github.com/zylisp/gisp/cmd/gisp/gisp.go
@@ -4,8 +4,8 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"github.com/rcarmo/gisp/generator"
-	"github.com/rcarmo/gisp/parser"
+	"github.com/zylisp/gisp/generator"
+	"github.com/zylisp/gisp/parser"
 	"go/ast"
 	"go/printer"
 	"go/token"
