+++ src/github.com/zylisp/gisp/core/number.go
