+++ src/github.com/zylisp/gisp/core/number_test.go
