+++ src/github.com/zylisp/gisp/generator/idents.go
@@ -1,7 +1,7 @@
 package generator
 
 import (
-	"github.com/rcarmo/gisp/parser"
+	"github.com/zylisp/gisp/parser"
 	"bytes"
 	"go/ast"
 	"regexp"
