+++ src/github.com/zylisp/gisp/generator/literals.go
