+++ src/github.com/zylisp/gisp/gitcommit.go
@@ -1,0 +1,3 @@
+package gisp
+
+func init() { GITLASTTAG = "0.5.0"; GITLASTCOMMIT = "1c6f191" }
